Split class for rewrite rule filtering.
